Give every language underlying and surface forms
Currently, the two translations are identical, except that underlying
forms look like /this/ and surface forms like <this>.

An NPC initiating a conversation with the adventurer in an language the
adventurer does not know will now say "Greetings" instead of "I don't
speak your language". They will still say the latter if the adventurer
initiates the conversation.

Units now use appropriate languages for conversations (currently the
underlying forms) instead of simply using the translation at a
hard-coded index.