Rewrite the phonemic inventory generator for speed
This approach generalizes the idea of dimensions. Instead of having one
dimension per child of the root node, there is one dimension per node.
Those dimensions are then iteratively joined in pairs to form a binary
tree of dimensions. A non-leaf dimension tries to get its values from
the cross product of its two subdimensions, prioritizing pairs similar
to those already chosen.

For example, the two subdimensions might correspond to place and manner.
The superdimension would start by picking a place and a manner, say
velar and fricative. For its next value, it would pick a new place or a
new manner, but not both; so it might produce an alveolar fricative. The
next value might then be an alveolar stop. The value after that, in that
case, would be a velar stop, to fill in the hole in the 2×2 grid. This
way, the phonemic inventory doesn't end up with a random hodge-podge of
phonemes like in Klingon. That may come later, but realistic inventories
are the priority.

The results are not great. There is still no way to ensure reasonable
numbers of both consonants and vowels. Secondary articulations are far
too common, though that might just be due to a poorly written raw file.
Still, it is good enough for now. And it is a lot faster.