Start syntax
This only models the phonological form. The logical form is not
important since DF already has a model of what an utterance means, which
I don't know how to modify anyway. It also assumes the language's
lexicon and parameters already exist; generating them will come later.

This model of syntax treats linguistic output as syntactic constituents
organized in a binary tree of nodes, each of which can have features.
These features can have values or not, and checking features leads to
movement of various kinds (raising, lowering, and local dislocation).
Morphemes can also fuse together irregularly. Bound morphemes are
supported with appropriate dummies.

The code is a big mess, but it nearly works. There are still some
problems with fusion, and lowering should really happen in a phase after
raising.