Rewrite phonemic inventory generation
The idea behind this attempt is that the dimension value picker, after
picking a round vowel (which is cross-linguistically common), should not
have a higher chance of picking a labialized consonant (which is not so
common) than it had before picking the vowel. To that end, it splits
dimension values into multiple copies whenever scalings apply to them,
and picks them independently.

It still doesn't work though. Dimensions seem so intuitively reasonable,
so why do they never work?

An easy but unsatisfying solution would be to manually specify valid
phonemes in the raws, instead of calculating them from the feature
geometry. That is only slightly better than the current DF system for
generating kobold and divine words.