Only apply scalings to applicable grid cells
It would apply a scaling to grid cell (i, j) if it was possible that at
least one cell in row i and at least one cell in column j might use the
scaling. This meant that if any scaling ever had a scalar of 0, the
phonemic inventory would always be empty.