Hack together a working inventory generator
It now generates acceptible inventories by encoding things in the raws
that I had previously tried to calculate from primitives. Dimensions are
now fully specified in the raws, using DIMENSION, DIMENSION_NODE, and
DIMENSION_END. Scalings can be explicitly associated with dimensions
using DIMENSION_SCALING. There is no restriction against inserting a
feature into the dimension tree multiple times.

That is all mainly just a change in the raw format. The only significant
code change is the new "peripheral" algorithm in `pick_from_grid`. This
algorithm, instead of choosing randomly from the cross product of the
two subdimensions, chooses alternatingly from the two subdimensions
directly. It is as if it is only picking from one row and one column of
the grid, which are like two sides of the periphery of the grid,
ignoring the more interesting cross-productive interior. It is really
only useful for the dimension which joins the vowels and the consonants.

Hopefully, this embarrassingly special-case algorithm can be generalized
and the dimension tree can be determined automatically. For now though,
this must suffice, lest I let the perfect be the enemy of the good.