Replace languages with lects
They are more or less the same thing, but Dwarf Fortress already uses
the word "language" for something else. Eventually lects will be used
for everything from language families to local slangs, so it will be a
more accurate term anyway.

Henceforth, language data will be stored in raw files, instead of in
persistent entries. This is more convenient both for internal use and
(eventually) for modding.

Note that lect generation does not work yet: it tries to load them from
files, but it can't. Perhaps the files have not yet been written.