Implement a proof-of-concept translation notation
The shorthand using `xp`, `x`, and `r` is not short enough. Verbose
boilerplate code is error-prone. The new format is shorter, reduces
rightwards drift, and is easily extensible. If a projection is split
(e.g. Infl) or some projections are rearranged, they need only be
modified in one function, instead of in each translation. The names of
the new desugaring functions begin with `ps_`, i.e. "phrase structure",
because they are ad hoc and non-binary, like phrase structure rules.
That's just a mnemonic, though; it's still X' underneath.