Insert rewritten announcements at correct indices
If multiple new conversation announcements appeared on the same
instant, each rewritten announcement would be inserted at the index of
its original. If rewriting changed the number of announcements, the
indices for all later announcements in the same instant would be wrong.
This is because it determined the index by searching the announcements
vector for a report with the right ID. However, if the number of
announcements changed, there could be two reports with the same ID. For
example, if the original vector had reports with IDs [6, 7, 8] and
report 6 were expanded to take two lines, the new vector would be [6, 7,
7, 8]. The first 7 is the second line of the rewritten form of report 6.
The second 7 is the original report 7. So far, so good. The rewriter
would then search for 7 and find the new 7, whereas it ought to rewrite
the old 7.

The solution is to search the announcements vector backwards.