Generate more varied phonemic inventories
Generating phonemic inventories the right way (i.e.
`get_dimension_values`) is slow and the results are always similar. This
is boring. The solution is to generate phonemes by randomly picking
phonetic symbols from the phonology. This is a hack on two counts.
First, symbols are part of the UI; they should have no effect on the
world. Second, the phonemes are generated without reference to each
other, so the resulting inventories make no sense.

I must admit, however, that the results are better gameplaywise, if not
linguistically. That is sad but not unexpected.